+++ eez_psu_sketch/ethernet.cpp
@@ -187,7 +187,7 @@ bool test() {
     return g_testResult != psu::TEST_FAILED;
 }
 
-void tick(unsigned long tick_usec) {
+void tick(uint32_t tick_usec) {
     if (g_testResult != psu::TEST_OK) {
         return;
     }
