+++ eez_psu_sketch/gui.h
@@ -26,7 +26,7 @@ namespace psu {
 namespace gui {
 
 void init();
-void tick(unsigned long tick_usec);
+void tick(uint32_t tick_usec);
 
 void refreshPage();
