+++ eez_psu_sketch/profile.h
@@ -83,7 +83,7 @@ struct Parameters {
     temperature::ProtectionConfiguration temp_prot[temp_sensor::MAX_NUM_TEMP_SENSORS];
 };
 
-void tick(unsigned long tick_usec);
+void tick(uint32_t tick_usec);
 
 void recallChannelsFromProfile(Parameters *profile);
 bool recallFromProfile(Parameters *profile);
