+++ eez_psu_sketch/serial_psu.h
@@ -27,7 +27,7 @@ namespace serial {
 extern scpi_t scpi_context;
 
 void init();
-void tick(unsigned long tick_usec);
+void tick(uint32_t tick_usec);
 
 }
 }
