+++ eez_psu_sketch/timer.cpp
@@ -22,7 +22,7 @@
 namespace eez {
 namespace psu {
 
-Interval::Interval(unsigned long interval_msec)
+Interval::Interval(uint32_t interval_msec)
 	: interval_usec(interval_msec * 1000L)
 {
 	reset();
@@ -32,8 +32,8 @@ void Interval::reset() {
 	next_tick_usec = micros() + interval_usec;
 }
 
-bool Interval::test(unsigned long tick_usec) {
-	long diff = tick_usec - next_tick_usec;
+bool Interval::test(uint32_t tick_usec) {
+	int32_t diff = tick_usec - next_tick_usec;
 	if (diff > 0) {
 		do {
 			next_tick_usec += interval_usec;
